Added lpm.lpm files (check description)
Introducing lpm.lpm! Make a new file in your root project called lpm.lpm, and make the whole file a directory to your main lua file! When you run `lpm run` or `lpm jitrun` with a lpm.lpm file in the folder you are running it in, it will attempt to run the lua file provided in lpm.lpm. No more typing out lua files again... And again... And again.