New install system and dependencies! (READ DESCRIPTION)
The new install system uses `curl` to install for more reliable and easier an easier to use installation process. It also provides more information such as how much time it will take to finish installing and how much time spent on it. Because of this new system, you can now add dependencies. Start out by making a `lpmpackage.json` file. Next put in it:
```
{
    "dependencies": ["package", "package2"]
}
```
Replace the packages with the packages you want to have lpm install with your package.